+++ .storybook/config.js
@@ -1,7 +1,7 @@
 import { configure, addDecorator, addParameters } from '@storybook/react';
-import centered from '@storybook/addon-centered/react';
 import { withPropsTable } from 'storybook-addon-react-docgen';
 import { create } from '@storybook/theming';
+import React from 'react'
 
 import '../src/components/style/index.css';
 
@@ -12,12 +12,15 @@ const theme = create({
 });
 
 addParameters({
+  backgrounds: [
+    { name: 'White', value: '#fff' },
+    { name: 'Navy', value: '#003366' },
+  ],
   options: {
     theme,
   },
 });
 
-addDecorator(centered);
 addDecorator(withPropsTable);
 
 // automatically import all files ending in *.stories.js
@@ -26,4 +29,4 @@ function loadStories() {
   req.keys().forEach((filename) => req(filename));
 }
 
-configure(loadStories, module);
+configure(require.context('../src/components', true, /\.stories\.(js|mdx)$/), module);