+++ src/components/modals/Dialog.jsx
@@ -38,8 +38,11 @@ const Dialog = (props) => {
     handleResize(); // needs to fire once immediately on mount
     // eslint-disable-next-line no-undef
     window.addEventListener('resize', handleResize);
-    // eslint-disable-next-line no-undef
-    return () => window.removeEventListener('resize', handleResize);
+    return () => {
+      noScroll.off();
+      // eslint-disable-next-line no-undef
+      window.removeEventListener('resize', handleResize);
+    };
   }, []);
 
   useLayoutEffect(() => {